+++ R/get_data.R
@@ -7,7 +7,8 @@
 #' @param subdomain One of the subdomains of hydroscope.gr
 #'
 #' @return If \code{subdomain} is one of \code{"kyy"} (Ministry of Environment
-#' and Energy), \code{"ypaat"} (Ministry of Rural Development and Food), or
+#' and Energy), \code{"ypaat"} (Ministry of Rural Development and Food),
+#' \code{"deh"} (Greek Public Power Corporation) or
 #' \code{"emy"} (National Meteorological Service), returns a tidy dataframe with
 #' stations' data from the corresponding database of hydroscope.gr. Otherwise
 #' gives an error message.
@@ -89,13 +90,15 @@
 #' \url{http://ypaat.hydroscope.gr}
 #' \item National Meteorological Service,
 #' \url{http://emy.hydroscope.gr}
+#' \item{Greek Public Power Corporation},
+#' \url{http://deh.hydroscope.gr}
 #'}
 #'
 #'
 #' @author Konstantinos Vantas, \email{kon.vantas@gmail.com}
 #' @import XML
 #' @export get_stations
-get_stations <- function(subdomain =  c("kyy", "ypaat", "emy")) {
+get_stations <- function(subdomain =  c("kyy", "ypaat", "emy", "deh")) {
 
   # match subdomain values
   subdomain <- match.arg(subdomain)
@@ -113,7 +116,7 @@ get_stations <- function(subdomain =  c("kyy", "ypaat", "emy")) {
 
     # get new stations from page
     new_stations <- tryCatch({
-      parseStations(url_page)
+      parse_stations(url_page)
     },
     error = function(e) {
       NULL
@@ -148,9 +151,10 @@ get_stations <- function(subdomain =  c("kyy", "ypaat", "emy")) {
 #'
 #' @return If \code{subdomain} is one of \code{"kyy"} (Ministry of Environment
 #' and Energy), \code{"ypaat"} (Ministry of Rural Development and Food),
-#' \code{"emy"} (National Meteorological Service) and stationID is not NULL,
-#' returns a dataframe with station's coordinates and elevation from the
-#' corresponding database of hydroscope.gr. Otherwise gives an error message.
+#' \code{"deh"} (Greek Public Power Corporation) or \code{"emy"} (National
+#' Meteorological Service), and stationID is not NULL, returns a dataframe with
+#' station's coordinates and elevation from the corresponding database of
+#' hydroscope.gr. Otherwise gives an error message.
 #'
 #'  If the station ID does not exist in the database, or the url from hydroscope
 #'  could not parsed, returns a dataframe with NA values.
@@ -190,11 +194,13 @@ get_stations <- function(subdomain =  c("kyy", "ypaat", "emy")) {
 #' \url{http://ypaat.hydroscope.gr}
 #' \item National Meteorological Service,
 #' \url{http://emy.hydroscope.gr}
+#' \item{Greek Public Power Corporation},
+#' \url{http://deh.hydroscope.gr}
 #'}
 #' @author Konstantinos Vantas, \email{kon.vantas@gmail.com}
 #' @import XML
 #' @export get_coords
-get_coords <- function(subdomain =  c("kyy", "ypaat", "emy"),
+get_coords <- function(subdomain =  c("kyy", "ypaat", "emy", "deh"),
                        stationID) {
 
   # check that stationID is given
@@ -209,37 +215,7 @@ get_coords <- function(subdomain =  c("kyy", "ypaat", "emy"),
 
   # Web scrapping --------------------------------------------------------------
   tryCatch({
-
-    # parse url
-    doc <- XML::htmlParse(url, encoding = "UTF8")
-
-    # get table nodes
-    path <- "/html/body/div[3]/div/div[1]/div/div[2]/div/div[2]/table"
-    tableNodes <- XML::getNodeSet(doc, path)
-
-    # check if table nodes is NULL
-    if (is.null(tableNodes)) stop("")
-
-    # read table
-    stat_table <- XML::readHTMLTable(tableNodes[[1]], header = FALSE,
-                                     stringsAsFactors = FALSE)
-    # get values from table
-    values <- stat_table$V2
-    names(values) <- stat_table$V1
-
-    # convert Coords to Lat and Long
-    Elevation <- as.numeric(values["Altitude"])
-    Coords <- as.character(values["Co-ordinates"])
-    tmp <- stringr::str_split(string = Coords, pattern = ",|\n",
-                              simplify = TRUE)
-    Lat <- as.numeric(tmp[1])
-    Long <- as.numeric(tmp[2])
-
-    # return results as a dataframe
-    data.frame(StationID = stationID,
-               Long = Long,
-               Lat = Lat,
-               Elevation = Elevation)
+    parse_coords(url, stationID)
   },
   error = function(e) {
     warning(paste0("Failed to parse url: ", url), call. = FALSE)
@@ -259,10 +235,11 @@ get_coords <- function(subdomain =  c("kyy", "ypaat", "emy"),
 #' @param stationID A station ID
 #'
 #' @return If \code{subdomain} is one of \code{"kyy"} (Ministry of Environment
-#' and Energy), \code{"ypaat"} (Ministry of Rural Development and Food) or
-#' \code{"emy"} (National Meteorological Service) and stationID is not NULL,
-#' returns a dataframe with time series data from the corresponding station and
-#' database of hydroscope.gr. Otherwise gives an error message.
+#' and Energy), \code{"ypaat"} (Ministry of Rural Development and Food),
+#' \code{"deh"} (Greek Public Power Corporation) or \code{"emy"} (National
+#' Meteorological Service), and stationID is not NULL, returns a dataframe with
+#' time series data from the corresponding station and database of
+#' hydroscope.gr. Otherwise gives an error message.
 #'
 #'  If the station ID does not exist in the database, or the url from hydroscope
 #'  could not parsed, returns a dataframe with NA values.
@@ -327,11 +304,13 @@ get_coords <- function(subdomain =  c("kyy", "ypaat", "emy"),
 #' \url{http://ypaat.hydroscope.gr}
 #' \item National Meteorological Service,
 #' \url{http://emy.hydroscope.gr}
+#' \item{Greek Public Power Corporation},
+#' \url{http://deh.hydroscope.gr}
 #'}
 #' @author Konstantinos Vantas, \email{kon.vantas@gmail.com}
 #' @import XML
 #' @export get_timeseries
-get_timeseries <- function(subdomain =  c("kyy", "ypaat", "emy"),
+get_timeseries <- function(subdomain =  c("kyy", "ypaat", "emy", "deh"),
                            stationID) {
 
   # check that stationID is given
@@ -344,42 +323,7 @@ get_timeseries <- function(subdomain =  c("kyy", "ypaat", "emy"),
 
   # Web scrapping --------------------------------------------------------------
   tryCatch({
-
-    # parse url
-    doc <- XML::htmlParse(url, encoding = "UTF8")
-
-    # get table nodes
-    tableNodes <- XML::getNodeSet(doc, "//*[@id=\"timeseries\"]")
-
-    # check if table nodes is NULL
-    if (is.null(tableNodes)) stop("")
-
-    # read table
-    ts_table <- XML::readHTMLTable(tableNodes[[1]], header = TRUE,
-                              stringsAsFactors = FALSE)
-    ts_table$StationID <- stationID
-
-    # check table dimensions
-    if (NROW(ts_table) == 0 | NCOL(ts_table) != 10) stop("")
-
-    # make names for table
-    names(ts_table) <- c("TimeSeriesID", "Name", "Variable", "TimeStep", "Unit",
-                    "Remarks", "Instrument", "StartDate", "EndDate",
-                    "StationID")
-
-    # Translations and transliterations ----------------------------------------
-    for (cname in c(names(ts_table))) {
-      ts_table[cname] <- greek2latin(ts_table[cname])
-    }
-
-    # translations
-    ts_table$Variable <- ts_variable(ts_table$Variable)
-    ts_table$TimeStep <- ts_timestep(ts_table$TimeStep)
-
-
-    # return dataframe
-    ts_table
-
+    parse_timeseries(url, stationID)
   },
   error = function(e) {
     warning(paste0("Failed to parse url: ", url), call. = FALSE)
@@ -397,10 +341,10 @@ get_timeseries <- function(subdomain =  c("kyy", "ypaat", "emy"),
 #' @param timeID A time series ID
 #'
 #' @return If \code{subdomain} is one of \code{"kyy"} (Ministry of Environment
-#' and Energy), \code{"ypaat"} (Ministry of Rural Development and Food) or
-#' \code{"emy"} (National Meteorological Service), and timeID is not NULL,
-#' returns a tidy dataframe with the time series values from the corresponding
-#' time series and database of hydroscope.gr. Otherwise gives an error message.
+#' and Energy) or \code{"ypaat"} (Ministry of Rural Development and Food),
+#' and timeID is not NULL, returns a tidy dataframe with the time series values
+#' from the corresponding database of hydroscope.gr. Otherwise gives an error
+#' message.
 #'
 #' If the time series ID does not exist in the database, or the url from
 #' Hydroscope could not parsed, returns a dataframe with NA values.
@@ -409,13 +353,17 @@ get_timeseries <- function(subdomain =  c("kyy", "ypaat", "emy"),
 #' \describe{
 #'     \item{Date}{The time series Dates (POSIXct)}
 #'     \item{Value}{The time series values (numeric)}
-#'     \item{Comment}{Comments from Hydroscope (character)}
+#'     \item{Comment}{Comments about the values (character)}
 #' }
 #'
 #' @note
-#' The subdomain "main" from Hydroscope is not used, because it uses different
-#' ID values for time series from the other subdomains and also because it does
-#' not contain the raw data of the time series.
+#' The subdomains \code{"deh"} (Greek Public Power Corporation) and \code{"emy"}
+#' (National Meteorological Service) are not used, because the data are not
+#' freely available.
+#'
+#' The subdomain "main", which also is not used, is a data
+#' aggregator from several databases. It uses different IDs for stations and
+#' time series.
 #'
 #' @examples
 #' # get time series 912 from the Greek Ministry of Environment and Energy
@@ -432,66 +380,29 @@ get_timeseries <- function(subdomain =  c("kyy", "ypaat", "emy"),
 #' \url{http://kyy.hydroscope.gr}
 #' \item Ministry of Rural Development and Food,
 #' \url{http://ypaat.hydroscope.gr}
-#' \item National Meteorological Service,
-#' \url{http://emy.hydroscope.gr}
 #'}
 #' @author Konstantinos Vantas, \email{kon.vantas@gmail.com}
 #' @import readr
 #' @export get_data
-get_data <- function(subdomain =  c("kyy", "ypaat", "emy"), timeID) {
+get_data <- function(subdomain =  c("kyy", "ypaat"), timeID) {
 
   # check that stationID is given
   if (is.null(timeID)) stop("argument \"timeID\" is missing")
 
-  # match subdomain values -----------------------------------------------------
+  # match subdomain values
   subdomain <- match.arg(subdomain)
-  url <- hydroscope_url(subdomain)
-  url <- paste0(url, "/timeseries/d/", timeID, "/download/")
+  h_url <- hydroscope_url(subdomain)
+  h_url <- paste0(h_url, "/api/tsdata/", timeID, "/")
 
-  # Download and convert hydroscope file to dataframe --------------------------
+  # get hydroscope file to dataframe
 
-  # create a temp file
-  tmp <- tempfile()
-    result<- tryCatch({
+  tryCatch({
+    enhy_get_txt(h_url)
+  },
+  error = function(e) {
+    # return NA values
+    warning(paste("Couldn't get time series' data from ", h_url), call. = FALSE)
+    dataNA()
+  })
 
-      # download file
-      dl_code <- utils::download.file(url = url, destfile = tmp)
-      if (dl_code != 0) stop("")
-
-      # Count the number of fields in each line of a file
-      cf <- readr::count_fields(tmp,
-                                tokenizer = readr::tokenizer_csv(),
-                                n_max = 50)
-
-      # check the number of columns in hts file
-      if (any(cf == 3)) {
-
-        # read timeseries data
-        tmFormat <- "%Y-%m-%d %H:%M"
-        result <- suppressWarnings(readr::read_csv(file = tmp,
-                                  skip =  sum(cf < 3),
-                                  col_names = c("Date", "Value", "Comment"),
-                                  col_types = list(
-                                    readr::col_datetime(format = tmFormat),
-                                    readr::col_double(),
-                                    readr::col_character())))
-
-        # remove NA Date values and return dataframe
-        result[!is.na(result$Date), ]
-
-      } else {
-        dataNA()
-      }
-    },
-    error = function(e) {
-      # return NA values
-      warning(paste("Couldn't get time series' data from ", url), call. = FALSE)
-      dataNA()
-
-    },
-    finally = {
-      # delete temp file
-      unlink(tmp)
-    })
-  return(result)
 }