+++ tests/testthat/test_get_functions.R
@@ -8,7 +8,7 @@ test_that("get_ functions give errors", {
   expect_error(get_coords(subdomain = "kyy"))
 
   expect_error(get_timeseries(subdomain = "none", stationID = 1))
-  expect_error(get_timeseries(subdomain = "kyy"))
+  expect_error(get_timeseries(subdomain = "deh"))
 
   expect_error(get_data(subdomain = "none", timeID = 1))
   expect_error(get_data(subdomain = "kyy"))
@@ -17,14 +17,18 @@ test_that("get_ functions give errors", {
 test_that("get_ functions give warnings", {
   expect_warning(get_coords("kyy", "a"))
   expect_warning(get_timeseries("emy", "a"))
-  expect_warning(get_data("emy", "a"))
+  expect_warning(get_data("deh", "a"))
 })
 
 test_that("get_ functions return dataframes", {
 
+  # return data franes
   expect_is(get_stations("ypaat"), "data.frame")
   expect_is(get_coords("kyy", 10004), "data.frame")
+  expect_is(get_coords("deh", 30022), "data.frame")
   expect_is(get_timeseries("kyy", 200206), "data.frame")
+  expect_is(get_timeseries("deh", 30022), "data.frame")
+
   expect_is(get_timeseries("emy", 20100), "data.frame")
   expect_is(get_data("kyy", 789), "data.frame")