+++ NEWS.md
@@ -1,7 +1,9 @@
 # hydroscoper 0.1.0.9000
 
-* Support hydroscope's multi-page front-end
-* Add station's and time series' data from all Hydroscope's domains
+* Use Enhydris API
+* Add Greek Public Power Corporation subdomain support
+* Add databases for all stations' and time series' data from all Hydroscope's 
+  domains
 
 # hydroscoper 0.1.0 (Release date: 2017-12-22)